requirements.txt
fastapi
uvicorn[standard]
pillow
opencv-python-headless
requests
numpy